Precision

+++ include/CMAPF/PCDefinitions.hpp
@@ -17,15 +17,15 @@ using namespace boost;
 
 	struct meta_data
 	{
-		std::pair <float, float> start;
-		std::pair <float, float> goal;
+		std::pair <double, double> start;
+		std::pair <double, double> goal;
 
 		std::vector <int> agent_list;
 
 		int task_id;
 
 		meta_data() {}
-		meta_data(std::pair <float, float> _start, std::pair <float, float> _goal, std::vector <int> _agent_list, int _task_id) : start(_start), goal(_goal), agent_list(_agent_list), task_id(_task_id) {}
+		meta_data(std::pair <double, double> _start, std::pair <double, double> _goal, std::vector <int> _agent_list, int _task_id) : start(_start), goal(_goal), agent_list(_agent_list), task_id(_task_id) {}
 	};
 
 	typedef property<meta_data_t, meta_data> MetaData;
